chore(T9751): delete trust-gate-adapter; inline core resolver in install.ts
Remove packages/caamp/src/core/skills/trust-gate-adapter.ts (229 LOC)
and inline the equivalent dynamic-import + graceful-degradation bridge
directly in packages/caamp/src/commands/skills/install.ts — the only
caller (verified via grep across all packages/).

caamp MUST NOT statically depend on @cleocode/core (the static graph is
core → caamp, never the reverse). The adapter file's sole purpose was to
wrap that constraint behind a typed API; with install.ts as the only
consumer, the helper folds in cleanly and the indirection disappears.

Inlined surface (local to install.ts):
  - resolveCore() — lazy dynamic-import + cached + stderr fallback
  - evaluateSkillTrustGate() — scanSkill + shouldAllowInstall passthrough
  - evaluateFederationGate()  — evaluateFederationInstallGate passthrough
  - recordSkillTrustBypass()  — recordTrustBypass passthrough

Behaviour is byte-equivalent to the old adapter: when core is absent
(standalone caamp install), the gates degrade to `allow` with a stderr
warning. All 1409 caamp tests pass. No CLI consumer breakage.

Wave D / Step 3 of T9740 (SG-CLEO-SKILLS cleanup).
